WIP: Mac: Set high physical DPI on HiDPI screens.
Use the two parameters that the platform screen
currently can change: logical DPI (fix to 72) and
physical size (calculate to get 72/144 physical DPI,
does not correspond to anything physical)

Add QWindow::virtualScreen(), which changes value
as the window is moved on to different (virtual)
screens. (QWindow::screen() is documented to not
change in this case.)

Add tests/manual/cocoa/hidpi.

Change-Id: Ica575d4d206ff577acb354bf82188d287476d7c9