WIP: HiDPI raster painting.
Create a 2X backing store and update [QNSView drawRect]
to draw it correctly.

Add a scaling factor to QImage, encode it as the
difference between the logical and physical DPI.

Attempt QPainter scaling based on the QImage DPI
scale:
- scale the redirection matrix
- scale the viewport
- scale the clip rect.

The idea is that drawing and clipping should happen
in logical point space, independently of resolution.

Painter scaling works well for tests/manual/cocoa/hidpi,
not so well for more complex examples.

Change-Id: I57a19e7c505fc13e12f3ed78800867940712802d