Clean up handling of SDKs on MacOS/iOS
Instead of setting -isysroot in both arch.test, compile.test, the iOS
mkspecs, and in sdk.prf, we now propgate the chosen SDK as the qmake
variable QMAKE_MAC_SDK, which is then handled exclusivly in sdk.prf.

The QMAKE_MAC_SDK variable, and -sdk argument to configure, is expected
to be of the short-form name, eg macosx or iphoneos, not a full path, as
that's what Xcode also expects. We take care of translating that into
a full path for -isysroot/-syslibroot in sdk.prf, using xcodebuild as
a helper.

Change-Id: I27de1ed318cd8bbfae2a95c33e23edb5187a0265
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>