iOS: Re-enable application bundles
The app bundles produces by the makefile generator may not work on iOS
yet, but neither does a non-bundle executable, and the primary way to
include application code in an iOS project is to link to a static lib
anyways.

We rely on Qt modules to have a .qmake.conf that loads(qt_build_config),
which will then set CONFIG += QTDIR_build, which we use as a signal that
we're building Qt itself. This fails for the config test, as they run
with -no-cache (.qmake.conf is aparently a cache file), and we can't
check the command line arguments like QT_BUILD_TREE, as the spec
is loaded before the arguments are evaluated. The workaround is
to look at the defaukt_pre we're processing and see if it comes
from the source directory or build directory. In the former case
we assume we're building Qt itself.

Later on we should try to find a non-hackish way to allow different
build config/options when building Qt and when building third party
apps, for example by splitting out qtbase's mkspecs/features into
one set that's global and installed, and one set that's only used
when building Qt, just like apps can have their own features.

Change-Id: Iace5052218f58291b648d663f89df4f299e52685
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>