QPA: update backingstore also when moving QWindows
The bug happens when moving native QMdiSubWindows (e.g a QGLWidget)
inside a QMdiArea. Since the views are native, QPA will assume
that it does not need to scroll the backingstore. But since the
backingstore can be shared among several windows stacked inside
the same hierarchy, this is still needed.

Change-Id: I35f4d4069a73ab6cd6d610e25c176c752789a927
QTBUG: 26162