iOS: Fix logic for choosing Xcode as output for client applications
We were trying to detect the usecase of client applications in default_pre,
which doesn't work as the project file hasn't been processed at that point
so TEMPLATE could be changed later on. We need to do the detection in
default_post instead.

Instead of setting the generator unconditionally we unset it in
default_pre so that we can detect if the user set it manually. This
means the user won't be able to inspect the MAKEFILE_GENERATOR variable
from the pro file, but this is less of a use-case then overriding the
generator from the command line or prooject file.

Change-Id: Idfe67380d2efa84a8aea57136c8a8933e961ab6e
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>