iOS: bugfix crash in QIOSInputContext
Add a QPointer to guard the GCD block in case the QIOSInputContext
gets deleted while the block is pending.
Note that setting m_focusView to 0 in the destructor is not safe
enough since the memory location can be overwritten before the block
gets executed. Retaining and releasing m_focusView around the block
is also not wanted, since we really want to close the keyboard even
if m_focusView changes while the block is pending (the old
m_focusView will then loose first responder status in favor of another
view).

Change-Id: I7546153adae310f2e458da9a483524d0a7b48656