qmacmime: bugfix unicode::canConvert
The UnicodeText convertor is a bit special since it can convert
to both utf8 and utf16 for text/plain. But since flavorFor("text/plain")
can only return one UTI, it returns utf16. And this means that
canConvert will return false if given the utf8 as argument.
On iOS this is often the only format available on the pasteboard, which
means that pasting text from another app into a Qt app will fail.

This patch will fix the text so that it returns true for both utf8 and
utf16.

Change-Id: Iac3fd8cd235fbad4e27c707c0d56a3c3d12848ee