iOS: remove unneeded bool block
We used m_ignoreKeyboardChanges to ensure that we didn't
scroll the screen up and down and emit signals when
we applied the resign/become first responder work-around
to update input traits (input method hints).

A better way is to instead delay processing the
keyboardDidChangeFrame notification. We might still
get try to scroll several times during a single event
loop callout, but since we now start the scroll from the
bounds of the display layer, this would effectively end
up as a no-op.

Change-Id: If39284053a1bd3b4f31343b0d32c0366c093869b