iOS: enable QUIView to be first responder without opening the keyboard
When a UIView becomes first responder, the keyboard will open as well
if the view implements the protocol UITextInput. It turns out that
we can override the static interface declaration runtime by
overriding - (BOOL)conformsToProtocol:(Protocol *)protocol
This means that we can choose to say that we implement the protocol
when, and only when, the keyboard should be visible. And further, we
can then allow QUIView to become first responder without opening
the keyboard. This will make things easier later when implementing
support for native menus that use the responder chain for
delivering menu actions.

Change-Id: I24577fe48d5ec752a72d10897c1dd64cc8d2bbfc