iOS: emit gestureTriggered from qiosinputcontext
qiosmenu needs to know when the gesture is triggered so it
can remove the event filter that override IM settings.
This is expecially important if the focus object that has the
menu 'attached' is root, since then resigning first responder
from the gesture will not cause the focus object to change.
In that case you always end up with a focus object that needs
the keyboard open because of the menu, which will stop the
gesture from working.

Change-Id: I860edbd555341365ebbc8fd39db69f5e34ea3744