iOS: clear object before opening a picker menu
Currently, when a picker menu open, it will not steal
focus, meaning that the menu will open while e.g a
line edit shows a blinking cursor. This is incorrect
since the line edit will not receive input from the keyboard.
To avoid this, we clear focus object before the menu opens.

This also has the added positive sideeffect that the menu
always attaches itself to the 'root' object of the window, instead
of whatever is the current focus object, which is more consistent.

Change-Id: I32ad628e2fcdce43cf85ff9d2128187bca46ee40