iOS: handle all directions when calculating positionFromPosition
The function should take UITextLayoutDirectionUp and
UITextLayoutDirectionDown into account.

Since IM in Qt does not support getting the position above
or below the current position, we just return the current
position, which is better than moving it left or right as we
do today.

We could add the missing IM API to Qt, but since we anyway
choose to handle the arrow keys explicitly as key events,
it would currently have no effect.

Change-Id: I4bcb9e2a00ab4e3d785058d7ff7f4855142dabbc