iOS: override setFrame in desktop view to always use the frame of the window
An assumption we do for the QIOSDesktopManagerView is that it should always
cover the whole UIWindow. To achieve that, we override setFrame to be stop
any attempt from UIKit to make our view smaller, e.g when the statusbar
changes height.

But the current attempt to calculate the frame goes wrong when the view has
a different orientation than the window (which is always in portrait). In that
case the view will have a transformation (rotation + translation). We
therefore need to map the frame of the window to the coordinate system of the
view to get the all the corners of the frame correct.

Task-number: QTBUG-47506
Change-Id: I388143f2cbb566541ffb1068443ce21e62ea2b42