iOS: send key events through QPA
So far we have chosen to send key events directly to the focus
object since we do that for IM events. But Qt expects key
events (control keys) to be sendt through QPA, like we do on
other platforms. While this might mean that some key events end up
somewhere else that at the focus object, it's still expected
behaviour, and needed if the short-cut propagation system is to
work.

Change-Id: I160bf3309572719eda352cdb11b46c4b5a455e0d