QLineEdit: add support for ImCursorPosition with position argument
Follow up f8dbed12, and implement support for IM queries that
can take an extra QVariant argument. We use this to implement
support for ImCursorPosition with a QPointF argument that
specifies that we should return the text position under
the given point rather than the current cursor pos (0bb645b).

Change-Id: I03ae7af698241f454d35dec84f248ecd5bb6ad51