configure/mkspec: check for xcodebuild rather than xcrun
xcrun seems to no longer be an included tool in Xcode8 beta (only a
tool in /usr/bin). So the tests we do for it in configure and
mkspec fails, which makes the whole Qt build fail.

This patch will change the tests to instead look for the presence
of xcodebuild.

Change-Id: I46ec73dc18165b9830e47d81fd75e2836204a30f