iOS, mkspec: remove faulty SDK check
We check if the SDK used by a project is less than 8.0
and error out if true. The intention of the test is to
avoid a situation where a project is buildt with an
SDK that is older than the one used to build Qt.

However, the test uses a hard coded value, and fails now that the
major iOS version has two digits. It hasn't been updated for
a while as well, since the test is currently checking for iOS 8.

So lets remove the check for now, until we can handle this in
a better way.

See QTBUG-37592.

Change-Id: I6106b9521b5d47d9906d4db30c2ffa21794bc307