iOS: don't override ImEnabled for menus
QIOSInputContext::inputMethodAccepted will print out a warning if
'QPlatformInputContext::inputMethodAccepted() does not match actual
focus object IM enablement'. And this happens whenever we use the
input panel for menus, since then we override ImEnabled in an
event filter to force the panel open.

A better solution than to 'fake' ImEnabled for menus (which can also
cause side effects elsewhere) is to check if the current focus
object has a custom input panel set when determining if the keyboard
should open.

Change-Id: If9f29cbceb964f6787223970a01743cc4ba48e2b