QWidgetTextControl/QLineEdit: support text position argument for ImCursorRectangle
By giving an extra argument telling the text position when querying for
ImCursorRectangle, we can ask for any rectangle in the text, not just
for the current cursor position.

We will utilize this from the iOS plugin to draw a temporary cursor in a
separate layer to greatly enhance performance when dragging the cursor
around together with a magnifier class (removing the need to do repaints
in the text control, which is really slow).

Change-Id: I1b1d6b429c6fa808f1593c06d074db2f5085e3d9